refactor(projects): pull out PC stuff
